#1694 - Improve annotation detail sidebar architecture and performance
- Pull out code for the upper layer selection frame (including the relation hint and forwarding checkbox) into a separate class/html
