+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/detail/AnnotationDetailEditorPanel.html
@@ -80,35 +80,9 @@
       </form>
       <div wicket:id="annotationFeatureForm" id="annotationFeatureForm" 
           class="annotatation-detail-panel flex-content flex-v-container flex-gutter">
-        <div wicket:id=deleteAnnotationDialog></div>
-        <div wicket:id=replaceAnnotationDialog></div>
-        <div wicket:id="layerContainer" class="card">
-          <!-- pb-1 compensates for the bottom padding of the form-group -->
-          <div class="card-body pb-1">
-            <div class="form-group form-row">
-              <label wicket:for="defaultAnnotationLayer" class="col-sm-3 col-form-label">
-                <wicket:label key="layer"/>
-              </label>
-              <div class="col-sm-9">
-                <select wicket:id="defaultAnnotationLayer" class="form-control" 
-                    data-container="body"></select>
-              </div>
-              <div class="col-sm-12 text-muted">
-                <small wicket:id="relationHint"></small>
-              </div>
-            </div>
-            <div class="form-group form-row" wicket:id="forwardAnnotationGroup">
-              <div class="col-sm-12">
-                <div class="form-check">
-                  <input wicket:id="forwardAnnotation" class="form-check-input" type="checkbox"/>
-                  <label wicket:for="forwardAnnotation" class="form-check-label">
-                    Forward annotation
-                  </label>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
+        <div wicket:id=deleteAnnotationDialog/>
+        <div wicket:id=replaceAnnotationDialog/>
+        <div wicket:id="layerContainer" class="card"/>
         <div class="flex-content panel card">
           <div class="card-header">
             Annotation
