+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/detail/LayerSelectionPanel.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<!--
+#Copyright 2020
+#Ubiquitous Knowledge Processing (UKP) Lab and FG Language Technology
+#Technische Universität Darmstadt
+#
+#Licensed under the Apache License, Version 2.0 (the "License");
+#you may not use this file except in compliance with the License.
+#You may obtain a copy of the License at
+# 
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+#Unless required by applicable law or agreed to in writing, software
+#distributed under the License is distributed on an "AS IS" BASIS,
+#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#See the License for the specific language governing permissions and
+#limitations under the License.
+-->
+<html xmlns:wicket="http://wicket.apache.org">
+<body>
+  <wicket:panel>
+    <!-- pb-1 compensates for the bottom padding of the form-group -->
+    <div class="card-body pb-1">
+      <div class="form-group form-row">
+        <label wicket:for="defaultAnnotationLayer" class="col-sm-3 col-form-label">
+          <wicket:label key="layer"/>
+        </label>
+        <div class="col-sm-9">
+          <select wicket:id="defaultAnnotationLayer" class="form-control" 
+              data-container="body"></select>
+        </div>
+        <div class="col-sm-12 text-muted">
+          <small wicket:id="relationHint"></small>
+        </div>
+      </div>
+      <div class="form-group form-row" wicket:id="forwardAnnotationGroup">
+        <div class="col-sm-12">
+          <div class="form-check">
+            <input wicket:id="forwardAnnotation" class="form-check-input" type="checkbox"/>
+            <label wicket:for="forwardAnnotation" class="form-check-label">
+              Forward annotation
+            </label>
+          </div>
+        </div>
+      </div>
+    </div>
+  </wicket:panel>
+</body>
+</html>
